EOS-18920: Patch-1: create separate validation modules and remove from configure_setup api (#1060)
* separate all validations from configure api and create separate modules for each Class


* comment to ignore codacy rule


* codacy fixes


* import path change
